move SysPathProvider to dedicated file to avoid cyclic dependency
