+++ src/databricks/labs/ucx/source_code/files.py
@@ -2,11 +2,10 @@ from __future__ import annotations  # for type hints
 
 import ast
 import logging
-import os
-import sys
 from pathlib import Path
-from collections.abc import Callable, Iterable
+from collections.abc import Callable
 
+from databricks.labs.ucx.source_code.syspath import SysPathProvider
 from databricks.sdk.service.workspace import Language
 
 from databricks.labs.ucx.source_code.languages import Languages
@@ -204,49 +203,3 @@ class LocalFileResolver(BaseDependencyResolver):
         if fullpath is not None:
             return Dependency(self._file_loader, fullpath)
         return super().resolve_import(name, problem_collector)
-
-
-class SysPathProvider:
-
-    @classmethod
-    def from_pathlike_string(cls, syspath: str):
-        paths = syspath.split(':')
-        return SysPathProvider([Path(path) for path in paths])
-
-    @classmethod
-    def from_sys_path(cls):
-        return SysPathProvider([Path(path) for path in sys.path])
-
-    def __init__(self, paths: list[Path]):
-        self._paths = paths
-        self._cwds = [Path(os.getcwd())]
-
-    def push_path(self, path: Path):
-        self._paths.insert(0, path)
-
-    def insert_path(self, index: int, path: Path):
-        self._paths.insert(index, path)
-
-    def remove_path(self, index: int):
-        del self._paths[index]
-
-    def pop_path(self) -> Path:
-        result = self._paths[0]
-        del self._paths[0]
-        return result
-
-    @property
-    def paths(self) -> Iterable[Path]:
-        yield from self._paths
-
-    def push_cwd(self, path: Path):
-        self._cwds.append(path)
-
-    def pop_cwd(self):
-        result = self._cwds[0]
-        del self._cwds[0]
-        return result
-
-    @property
-    def cwd(self):
-        return self._cwds[-1] if len(self._cwds) > 0 else Path(os.getcwd())
