+++ tests/unit/source_code/test_s3fs.py
@@ -8,7 +8,8 @@ from databricks.labs.ucx.source_code.graph import (
     DependencyProblem,
     DependencyGraphBuilder,
 )
-from databricks.labs.ucx.source_code.files import FileLoader, LocalFileResolver, SysPathProvider
+from databricks.labs.ucx.source_code.files import FileLoader, LocalFileResolver
+from databricks.labs.ucx.source_code.syspath import SysPathProvider
 from databricks.labs.ucx.source_code.whitelist import WhitelistResolver, Whitelist
 from tests.unit import (
     _load_sources,
